[refactor] move files to be under security package
