+++ src/main/java/org/example/travelexpertwebbackend/security/JwtAuthenticationFilter.java
@@ -1,9 +1,10 @@
-package org.example.travelexpertwebbackend.service.auth;
+package org.example.travelexpertwebbackend.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.example.travelexpertwebbackend.service.auth.UserService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
